refactor: isolate UI components from router component, add to storyboard
