+++ src/components/CreatePassword/CreatePasswordForm.stories.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+import { themeDecorator } from "../../storybookUtils";
+import { CreatePasswordForm } from "./CreatePasswordForm";
+
+export const Standard: React.FC = () => <CreatePasswordForm />;
+
+export default {
+  title: "Components/CreatePasswordForm",
+  decorators: [themeDecorator()],
+  component: CreatePasswordForm,
+};
